boot: Allow the kernel to run at EL2
At the moment we always drop to EL1 to run the kernel, completely
ignoring any Hypervisor functionality offered by the CPU. Given that
we don't yet support Bhyve we have no need for EL2 at present. However,
our boot process does not survive booting on systems that offer
Virtualisation Host Extensions (VHE). There are many reasons for this,
including sign-extension bugs and the fact the some register bits are
different when VHE is present. This was uncovered when booting on the
Qemu SBSA reference VM, which offers a Neoverse-N2 CPU by default.

On real hardware, the Raspberry Pi 4 offers EL2, but without VHE, as
does Qemu virt with the `virtualization=on` machine flag. In these
cases we need to install a virtualisation trap handler to EL2 prior
to dropping to EL1. We install a stub trap table (much like FreeBSD),
allowing us to replace the table with a full trap table later. Ampere
Altra and Altra Max systems are similar to Pi4 in this regard.

On modern hardware with VHE (Qemu sbsa-ref, AmpereOne, Grace) we need
to run the kernel at EL2, reducing the number of context switches
needed for Hypervisor operations. In this setup we do not install a
stub trap table, as we run at EL2 (so we can install a table at any
time).

So, we have:
* Boot at EL1, no Hypervisor extensions.
* Boot at EL2, no VHE, install stubs, drop to EL1.
* Boot at EL2, VHE, configure unified EL12 and stay at EL2.

Worth noting is that due to the way we boot, we replace the stub
trap table while starting the kernel. This is because the code
executed by `loader(7)` is backed by different physical pages than
the running kernel and because booting secondary processors uses
temporary boot pages.

The side-effect of this change is that we can boot to the end of
the loader on Qemu SBSA, which implements Neoverse-N2 with VHE. This
has also been tested to boot into the kernel (but only the early
kernel, as ACPI support is still a work in progress for the kernel).

The revised startup code has heavy reference to FreeBSD, with the
intent of adding Bhyve/aarch64 support in the future (once that is
fully merged into FreeBSD and stable).