aarch64: hat: Don't clean dcache for non-memory pages
The dcache cleaning implementation assumes that the referenced PFN can
be found mapped in segkpm, but we're actively removing device mappings
from that region. Attempting to clean an unmapped range in the KVA
traps, causing boot to fail when the boot UART (or framebuffer) is
remapped, which triggers this cache-cleaning for a non-existent
mapping.

As with all fun-to-debug bugs this only showed up on real hardware,
which was rpi4 in my case.