loader: Add the ARMH,sbbr implementation architecture
ACPI-based aarch64 systems will broadly comply with the Arm
SystemReady-SR or Systemready-ES standards families (and might comply
with the SystemReady-IL standards).

Since illumos ACPI support is intended to be a fairly permissive
interpretation of the Server Base Boot Requirements (DEN0044 [1]),
the implementation architecture (platform) will be named `ARMH,sbbr`.

There's no kernel for `ARMH,sbbr` at the moment, but this seems like
an opportune time to recognise what it will be and add support to the
loader.

[1] https://developer.arm.com/documentation/den0044/latest/