armv8: Import gfx_private for aarch64
This is a lightly edited copy of gfx_private from
`i86pc/io/gfx_private`.

VGA support has been removed, as that is `i86pc`-specific. Additionally,
PCI support has been removed and/or disabled as it uses older-style
interfaces to PCI configuration space. This part of the change may
need to be reworked later (in addition to tracking which PCI device
is represented by the UEFI framebuffer).

Eventually the `i86pc` version of this code could probably be moved
to `uts/common` and lightly reworked to make VGA and PCI optional (or
platform-specific).