armv8: efifb for aarch64
Add a simple efifb driver. This is a clone of the vgatext driver from
`uts/intel`, warts and all.

Eventually, after moving `gfx_private` to `uts/common`, we could
also move `vgatext` to common and use the same code on both platforms,
just leaving the VGA portions out on platforms that don't do that.