viommionex: Defer to parent bus where appropriate
The `viommionex` driver can't have knowledge of what any parent busses
may need, but needs to gracefully handle differences in bus-specific
address cells handling. Until now the driver has managed to get away
with simply using the standard `bus_map` function as well as standard
parent private data on any child devices. With the introduction of
ACPI, which has it's own `regs` format, the existing use of the standard
mechanisms no longer works.

Pass child device initialiasation and uninitialisation to the bus parent
to allow the parent to fill the appropriate structures as needed, then
override the `bus_map` function to call the parent's `bus_map` when that
is present, falling back to the existing standard function when the
parent does not provide a `bus_map`.

This change allows `viommionex` to be a child of `rootnex` (as it is
today), `simple-bus` (should already work) and `acpinex`.