pci_boot: aggressively reprogram PCI buses
Change tack, with inspiration from Oxide, and always reprogram the PCI buses.

Our hope that we could trust firmware to do this turns out to be regrettably
misplaced.  Even using the same firmware to do the same things, the bus on
RaspberryPi 4 is disabled, unlike with standards complaint ecam devices.

So, instead, we follow the lead of Oxide and aggressively reprogram.  We do
this more thoroughly than Oxide, and we remove the (now dead) code to accept
firmware programming in the hopes of a more simple and predictable result.

Portions Contributed By: Andy Fiddaman <illumos@fiddaman.net>