bootdev: ship on arm, it's absolutely required but only rarely
The mechanism we require this through is a bit weird.

The short answer is that we define modstubs as being within this module, we
thus commit to providing the module and providing concrete versions.  oops.

The long answer includes why this seems to only crash in a minority of cases,
we never call these stubbed functions that I can find _except_ when we are
rebooting, and even then perhaps only when rebooting after a forced boot
archive rebuild.

Provide what intel does, in the same ways.

I've "tested this" by managing to trigger the problem circumstances with
breakpoints in these functions, to see that they were
called (`i_convert_boot_device_name` is, twice), and then continuing from
them.  They're called so deep into rebooting it is hard to say really whether
the only one that is called -- a **strdup(3C)** in sheep's clothing -- does
the right thing.  It does work though.

I'm not sure if there is any benefit to **misc/bootdev** staying a module when
nobody overrides anything it does on any basis whatever.