pcie: remove warning about `pcie_get_rc_dip`
We changed this such that it would return `dip` if `dip` was itself the root
complex, and left a note in case that broke anything.

I've been through callers and I don't think that's a problem.  The problem I
anticipate would be something where we do `pcie_get_rc_dip(dip) != NULL`.