gicthree: Propagate redistributor configuration at CPU startup
We have been assuming that a redistributor for an inactive CPU can still be
programmed via the register interface. It turns out that some CPUs ignore
writes to distributors that have not been woken. This is the case on both
Qemu sbsa-ref and Ampere Altra Max.

Operating under the above incorrect assumption leads to failure to boot most
application processors through general failure to IPI other CPUs, since those
CPUs are not configured to receive SGIs. The flow of events is:
  CPU0 starts CPU1
  CPU1 registers, calling pause_cpus
    pause_cpus IPIs other CPUs to quiesce them
    CPU0 receives the IPI and quiesces
  CPU1 completes registration, calling start_cpus
    start_cpus IPIs other CPUs to awaken them
    CPU0 receives the IPI and awakens
  CPU0 starts CPU2
  CPU2 registers, calling pause_cpus
    pause_cpus IPIs other CPUs to quiesce them
    CPU0 receives the IPI and quiesces
    CPU1 does NOT receive the IPI
  CPU2 spins, waiting for CPU1 to quiesce
  ... cpu_pause never completes...

This change programs initial redistributor settings from defaults when
configuring the boot processor, then programs redistributors for
application processors from the current state of the boot processor at the time
the application processor starts, propagating SGI and PPI configuration to the
application processor.

This is, morally, what happens in the gictwo (the GICv2 driver), but
without having to manually track the register state for SGI and PPI
interrupts.