kobj_isa: aarch64: Rework instruction cache flushing
Avoid the CTR_ macros from sys/controlregs.h, as those are backed by
bitx64, which lives in genunix and may not be resolved by the time
this code is executed. Even if bitx64 has been resolved, this code
could (and does) blow up on systems with a sufficiently large
instruction prefetch buffer, such as Altra Max.